+++ src/components/mate/MateCard.tsx
